+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.21.1](https://github.com/Greenstand/treetracker-admin-client/compare/v1.21.0...v1.21.1) (2021-09-25)
+
+
+### Bug Fixes
+
+* add uuid to capture detail dialog and top filter ([#166](https://github.com/Greenstand/treetracker-admin-client/issues/166)) ([5771c73](https://github.com/Greenstand/treetracker-admin-client/commit/5771c735103c3dff78d957b1b6f40a8c59ef265d))
+
+# [1.21.0](https://github.com/Greenstand/treetracker-admin-client/compare/v1.20.3...v1.21.0) (2021-09-24)
+
+
+### Features
+
+* **planter filter:** include partial matches in results ([#165](https://github.com/Greenstand/treetracker-admin-client/issues/165)) ([90d493e](https://github.com/Greenstand/treetracker-admin-client/commit/90d493e404e882b568a78d0d957d7c23a7cdae54))
+
+## [1.20.3](https://github.com/Greenstand/treetracker-admin-client/compare/v1.20.2...v1.20.3) (2021-09-24)
+
+
+### Bug Fixes
+
+* planter filter to allow filtering parent orgs by name and all orgs by id ([#164](https://github.com/Greenstand/treetracker-admin-client/issues/164)) ([ee3b6f0](https://github.com/Greenstand/treetracker-admin-client/commit/ee3b6f08103f7d0fc706a390efa012ba66af0ad7))
+
+## [1.20.2](https://github.com/Greenstand/treetracker-admin-client/compare/v1.20.1...v1.20.2) (2021-09-22)
+
+
+### Bug Fixes
+
+* replace logo png with svg ([#157](https://github.com/Greenstand/treetracker-admin-client/issues/157)) ([52e38bf](https://github.com/Greenstand/treetracker-admin-client/commit/52e38bf71232947602048b7412df0e6f4a786ed0))
+
 ## [1.20.1](https://github.com/Greenstand/treetracker-admin-client/compare/v1.20.0...v1.20.1) (2021-09-17)